requirements.txt
setuptools==36.5.0
python_neutronclient==6.7.0
tabulate==0.8.2
requests==2.18.4
python_heatclient==1.10.0
docopt==0.6.2
coloredlogs==9.0
dill==0.2.7.1
keystoneauth1==3.3.0
click==6.7
python-dotenv==0.7.1
GitPython==2.1.8
python_keystoneclient==3.15.0
python_novaclient==10.1.0
PyYAML==3.12
